% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{browse_dataset}
\alias{browse_dataset}
\title{Browse the dataset in your browser}
\usage{
browse_dataset(con, dataset_id)
}
\arguments{
\item{con}{A \linkS4class{DomoConnection} object}

\item{dataset_id}{The dataset's ID as a string}
}
\description{
Browse the dataset in your browser
}
