% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\name{dbConnect,DomoDriver-method}
\alias{dbConnect,DomoDriver-method}
\title{Connect to a Domo instance}
\usage{
\S4method{dbConnect}{DomoDriver}(
  drv,
  client_id = domo_client_id(),
  client_secret = domo_client_secret(),
  ...
)
}
\arguments{
\item{drv}{an object that inherits from \linkS4class{DBIDriver},
or an existing \linkS4class{DBIConnection}
object (in order to clone an existing connection).}

\item{client_id}{OAuth 2.0 client ID as a string.}

\item{client_secret}{OAuth 2.0 client secret as a string.}

\item{...}{These dots are for future extensions and must be empty.}
}
\description{
Connect to a Domo instance
}
\details{
By default the \code{client_id} and \code{client_secret} are read in from environment
variables \code{DOMO_CLIENT_ID} and \code{DOMO_CLIENT_SECRET}, respectively. To obtain
the credentials, login to your Domo instance and visit
\href{https://developer.domo.com/new-client}{Domo's developer portal}. The client
must have the "Data" application scope.
}
