% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.R
\name{dbFetch,DomoResult-method}
\alias{dbFetch,DomoResult-method}
\title{Retrieve records from Domo query}
\usage{
\S4method{dbFetch}{DomoResult}(res, n = -1, ...)
}
\description{
Retrieve records from Domo query
}
