% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{dbSendQuery,DomoConnection,character-method}
\alias{dbSendQuery,DomoConnection,character-method}
\title{Query Domo dataset}
\usage{
\S4method{dbSendQuery}{DomoConnection,character}(conn, statement, dataset_id, ...)
}
\arguments{
\item{conn}{A \linkS4class{DBIConnection} object, as returned by
\code{\link[DBI:dbConnect]{dbConnect()}}.}

\item{statement}{a character string containing SQL.}

\item{dataset_id}{The dataset ID (e.g. 77194824-a5c1-11ea-9600-0116159276e2).}

\item{...}{Other parameters passed on to methods.}
}
\description{
Query Domo dataset
}
