% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domo-package.R
\docType{package}
\name{domo-package}
\alias{domo}
\alias{domo-package}
\title{domo: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Shun Sambongi \email{shundev3@gmail.com}

}
\keyword{internal}
