% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{domo-tables}
\alias{domo-tables}
\alias{dbCreateTable,DomoConnection-method}
\alias{dbWriteTable,DomoConnection,character-method}
\alias{dbAppendTable,DomoConnection,character-method}
\alias{dbReadTable,DomoConnection,character-method}
\alias{dbRemoveTable,DomoConnection,character-method}
\alias{dbExistsTable,DomoConnection,character-method}
\alias{dbListTables,DomoConnection-method}
\alias{dbListFields,DomoConnection,character-method}
\title{Convenience functions for reading/writing Domo datasets}
\usage{
\S4method{dbCreateTable}{DomoConnection}(conn, name, fields, ..., row.names = NULL, temporary = FALSE)

\S4method{dbWriteTable}{DomoConnection,character}(
  conn,
  name,
  value,
  ...,
  overwrite = FALSE,
  append = FALSE,
  stream = TRUE
)

\S4method{dbAppendTable}{DomoConnection,character}(conn, name, value, ..., stream = TRUE, row.names = NULL)

\S4method{dbReadTable}{DomoConnection,character}(conn, name, ...)

\S4method{dbRemoveTable}{DomoConnection,character}(conn, name, ...)

\S4method{dbExistsTable}{DomoConnection,character}(conn, name, ...)

\S4method{dbListTables}{DomoConnection}(conn, ...)

\S4method{dbListFields}{DomoConnection,character}(conn, name, ...)
}
\arguments{
\item{conn}{A \linkS4class{DBIConnection} object, as returned by
\code{\link[DBI:dbConnect]{dbConnect()}}.}

\item{name}{\itemize{
\item For \code{dbCreateTable}: the new dataset's name
\item For everything else: an existing dataset's ID
}}

\item{fields}{Either a character vector or a data frame.

A named character vector: Names are column names, values are types.
Names are escaped with \code{\link[DBI:dbQuoteIdentifier]{dbQuoteIdentifier()}}.
Field types are unescaped.

A data frame: field types are generated using
\code{\link[DBI:dbDataType]{dbDataType()}}.}

\item{...}{Other arguments used by individual methods.}

\item{row.names}{Must be \code{NULL}.}

\item{temporary}{Must be \code{FALSE}. Domo cannot create temporary datasets.}

\item{value}{a \link{data.frame} (or coercible to data.frame).}

\item{overwrite}{Allow overwriting the destination dataset. Cannot be \code{TRUE}
if \code{append} is also \code{TRUE}.}

\item{append}{Allow appending to the destination table. Cannot be \code{TRUE}
if \code{overwrite} is also \code{TRUE}.}

\item{stream}{If \code{TRUE} use
\href{https://developer.domo.com/docs/streams-api-reference/streams}{Stream API},
otherwise use \href{https://developer.domo.com/docs/dataset-api-reference/dataset}{Dataset API}.}
}
\description{
Convenience functions for reading/writing Domo datasets
}
\examples{
\dontrun{
library(DBI)
con <- dbConnect(domo::domo())
dbListTables(con)
dataset_id <- dbCreateTable(con, "mtcars", mtcars)
dbWriteTable(con, dataset_id, mtcars, overwrite = TRUE)
dbReadTable(con, dataset_id)

dbExistsTable(con, dataset_id)

dbDisconnect(con)
}

}
