% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.R
\docType{class}
\name{Domo}
\alias{Domo}
\alias{dbConnect,DomoDriver-method}
\title{Domo connection class}
\usage{
\S4method{dbConnect}{DomoDriver}(
  drv,
  client_id = domo_client_id(),
  client_secret = domo_client_secret(),
  scope = "data"
)
}
\arguments{
\item{drv}{An object created by \code{\link[=domo]{domo()}}}
}
\description{
Domo connection class
}
\keyword{internal}
