% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.R, R/table.R
\docType{class}
\name{DomoConnection-class}
\alias{DomoConnection-class}
\alias{show,DomoConnection-method}
\alias{dbDisconnect,DomoConnection-method}
\alias{dbIsValid,DomoConnection-method}
\alias{dbDataType,DomoConnection-method}
\alias{dbSendQuery,DomoConnection,character-method}
\alias{dbCreateTable,DomoConnection-method}
\alias{dbWriteTable,DomoConnection,character-method}
\alias{dbReadTable,DomoConnection,character-method}
\alias{dbRemoveTable,DomoConnection,character-method}
\alias{dbExistsTable,DomoConnection,character-method}
\alias{dbListTables,DomoConnection-method}
\alias{dbListFields,DomoConnection,character-method}
\title{Class DomoConnection (and methods)}
\usage{
\S4method{show}{DomoConnection}(object)

\S4method{dbDisconnect}{DomoConnection}(conn, ...)

\S4method{dbIsValid}{DomoConnection}(dbObj, ...)

\S4method{dbDataType}{DomoConnection}(dbObj, obj, ...)

\S4method{dbSendQuery}{DomoConnection,character}(conn, statement, dataset_id, ...)

\S4method{dbCreateTable}{DomoConnection}(conn, name, fields, ..., row.names = NULL, temporary = FALSE)

\S4method{dbWriteTable}{DomoConnection,character}(
  conn,
  name,
  value,
  ...,
  overwrite = FALSE,
  append = FALSE,
  stream = TRUE
)

\S4method{dbReadTable}{DomoConnection,character}(conn, name, ...)

\S4method{dbRemoveTable}{DomoConnection,character}(conn, name, ...)

\S4method{dbExistsTable}{DomoConnection,character}(conn, name, ...)

\S4method{dbListTables}{DomoConnection}(conn, ...)

\S4method{dbListFields}{DomoConnection,character}(conn, name, ...)
}
\description{
DomoConnection objects are created by passing \code{\link[=domo]{domo()}} as the first argument
to \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.
They are a subclass of the \linkS4class{DBIConnection} class.
}
\keyword{internal}
