% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\docType{class}
\name{DomoConnection}
\alias{DomoConnection}
\alias{DomoConnection-class}
\alias{show,DomoConnection-method}
\alias{dbIsValid,DomoConnection-method}
\alias{dbDisconnect,DomoConnection-method}
\alias{dbDataType,DomoConnection,ANY-method}
\alias{dbDataType,DomoConnection,data.frame-method}
\alias{dbGetInfo,DomoConnection-method}
\title{Domo Connection Methods}
\usage{
\S4method{show}{DomoConnection}(object)

\S4method{dbIsValid}{DomoConnection}(dbObj, ...)

\S4method{dbDisconnect}{DomoConnection}(conn, ...)

\S4method{dbDataType}{DomoConnection,ANY}(dbObj, obj, ...)

\S4method{dbDataType}{DomoConnection,data.frame}(dbObj, obj, ...)

\S4method{dbGetInfo}{DomoConnection}(dbObj, ...)
}
\arguments{
\item{object}{Any R object}

\item{dbObj}{An object inheriting from \linkS4class{DBIObject},
i.e. \linkS4class{DBIDriver}, \linkS4class{DBIConnection},
or a \linkS4class{DBIResult}}

\item{...}{Other arguments to methods.}

\item{conn}{A \linkS4class{DBIConnection} object, as returned by
\code{\link[DBI:dbConnect]{dbConnect()}}.}

\item{obj}{An R object whose SQL type we want to determine.}
}
\description{
Implementations of pure virtual functions defined in the \code{DBI} package
for DomoConnection objects.
}
