% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.R
\docType{class}
\name{DomoDriver-class}
\alias{DomoDriver-class}
\alias{dbUnloadDriver,DomoDriver-method}
\title{Driver for Domo}
\usage{
\S4method{dbUnloadDriver}{DomoDriver}(drv, ...)
}
\description{
Driver for Domo
}
\keyword{internal}
