% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\docType{class}
\name{DomoDriver}
\alias{DomoDriver}
\alias{DomoDriver-class}
\alias{show,DomoDriver-method}
\alias{dbDataType,DomoDriver,ANY-method}
\alias{dbDataType,DomoDriver,data.frame-method}
\alias{dbIsValid,DomoDriver-method}
\alias{dbGetInfo,DomoDriver-method}
\title{Domo Driver Methods}
\usage{
\S4method{show}{DomoDriver}(object)

\S4method{dbDataType}{DomoDriver,ANY}(dbObj, obj, ...)

\S4method{dbDataType}{DomoDriver,data.frame}(dbObj, obj, ...)

\S4method{dbIsValid}{DomoDriver}(dbObj, ...)

\S4method{dbGetInfo}{DomoDriver}(dbObj, ...)
}
\arguments{
\item{object}{Any R object}

\item{dbObj}{A object inheriting from \linkS4class{DBIDriver}
or \linkS4class{DBIConnection}}

\item{obj}{An R object whose SQL type we want to determine.}

\item{...}{Other arguments passed on to methods.}
}
\description{
Implementation of pure virtual functions defined in the \code{DBI} package for
DomoDriver objects.
}
