% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\docType{methods}
\name{DomoResult}
\alias{DomoResult}
\alias{DomoResult-class}
\alias{show,DomoResult-method}
\alias{dbIsValid,DomoResult-method}
\alias{dbBind,DomoResult-method}
\alias{dbFetch,DomoResult-method}
\alias{dbGetRowCount,DomoResult-method}
\alias{dbGetRowsAffected,DomoResult-method}
\alias{dbGetStatement,DomoResult-method}
\alias{dbColumnInfo,DomoResult-method}
\alias{dbClearResult,DomoResult-method}
\alias{dbHasCompleted,DomoResult-method}
\title{Domo Result Methods}
\usage{
\S4method{show}{DomoResult}(object)

\S4method{dbIsValid}{DomoResult}(dbObj, ...)

\S4method{dbBind}{DomoResult}(res, params, ...)

\S4method{dbFetch}{DomoResult}(res, n = -1, ...)

\S4method{dbGetRowCount}{DomoResult}(res, ...)

\S4method{dbGetRowsAffected}{DomoResult}(res, ...)

\S4method{dbGetStatement}{DomoResult}(res, ...)

\S4method{dbColumnInfo}{DomoResult}(res, ...)

\S4method{dbClearResult}{DomoResult}(res, ...)

\S4method{dbHasCompleted}{DomoResult}(res, ...)
}
\arguments{
\item{object}{Any R object}

\item{dbObj}{An object inheriting from \linkS4class{DBIObject},
i.e. \linkS4class{DBIDriver}, \linkS4class{DBIConnection},
or a \linkS4class{DBIResult}}

\item{...}{Other arguments to methods.}

\item{res}{An object inheriting from \linkS4class{DBIResult}.}

\item{params}{A list of bindings, named or unnamed.}

\item{n}{maximum number of records to retrieve per fetch. Use \code{n = -1}
or \code{n = Inf}
to retrieve all pending records.  Some implementations may recognize other
special values.}
}
\description{
Implementations of pure virtual functions defined in the \code{DBI} package for
DomoResult objects.
}
\examples{
\dontrun{
library(DBI)
con <- dbConnect(domo::domo())
dataset_id <- "77194824-a5c1-11ea-9600-0116159276e2"
res <- dbSendQuery(con, "SELECT mpg FROM mtcars WHERE mpg > ?", dataset_id)
dbBind(res, list(20))
dbHasCompleted(res)
dbGetRowCount(res)

dbFetch(res, n = 1)
dbHasCompleted(res)
dbGetRowCount(res)

dbFetch(res)
dbHasCompleted(res)
dbGetRowCount(res)

dbClearResult(res)
dbDisconnect(con)
}
}
