% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.R
\docType{class}
\name{DomoTblConnection-class}
\alias{DomoTblConnection-class}
\alias{dbSendQuery,DomoTblConnection,character-method}
\title{DomoTblConnection class}
\usage{
\S4method{dbSendQuery}{DomoTblConnection,character}(conn, statement, ...)
}
\description{
DomoTblConnection class
}
\keyword{internal}
