% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.R
\name{dbSendQuery,DomoConnection,character-method}
\alias{dbSendQuery,DomoConnection,character-method}
\title{Send a query to Domo}
\usage{
\S4method{dbSendQuery}{DomoConnection,character}(conn, statement, dataset_id, ...)
}
\description{
Send a query to Domo
}
