% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domo-package.R
\docType{package}
\name{domo-package}
\title{domo: DBI for DOMO}
\description{
This package provides a DBI interface for DOMO datasets.
}
\author{
\strong{Maintainer}: Shun Sambongi \email{shundev3@gmail.com}

}
\keyword{internal}
