% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domo-package.R
\docType{package}
\name{domo-package}
\title{domo: DBI for DOMO}
\description{
This package provides a DBI interface for DOMO
    datasets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://shunsambongi.github.io/domo}
  \item \url{https://github.com/shunsambongi/domo}
  \item Report bugs at \url{https://github.com/shunsambongi/domo/issues}
}

}
\author{
\strong{Maintainer}: Shun Sambongi \email{shundev3@gmail.com}

}
\keyword{internal}
