% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.R
\name{domo}
\alias{domo}
\alias{dbConnect,DomoDriver-method}
\title{Connect to a Domo instance}
\usage{
domo()

\S4method{dbConnect}{DomoDriver}(
  drv,
  client_id = domo_client_id(),
  client_secret = domo_client_secret(),
  scope = "data"
)
}
\arguments{
\item{drv}{An object created by \code{\link[=domo]{domo()}}}

\item{client_id}{OAuth 2.0 client ID as a string}

\item{client_secret}{OAuth 2.0 client secret as a string}

\item{scope}{A character vector of OAuth 2.0 scopes}
}
\value{
\itemize{
\item For \code{domo()}: An object of class \linkS4class{DomoDriver}
\item For \code{dbConnect()}: An object of class \linkS4class{DomoConnection}
}
}
\description{
\code{domo()} and \code{\link[DBI:dbConnect]{DBI::dbConnect()}} allow you to connect to a Domo instance, to
read and write from Domo datasets.
}
