% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\docType{class}
\name{DomoResult-class}
\alias{DomoResult-class}
\alias{show,DomoResult-method}
\alias{dbIsValid,DomoResult-method}
\alias{dbBind,DomoResult-method}
\alias{dbFetch,DomoResult-method}
\alias{dbGetRowCount,DomoResult-method}
\alias{dbGetStatement,DomoResult-method}
\alias{dbColumnInfo,DomoResult-method}
\alias{dbClearResult,DomoResult-method}
\alias{dbHasCompleted,DomoResult-method}
\title{Domo results class}
\usage{
\S4method{show}{DomoResult}(object)

\S4method{dbIsValid}{DomoResult}(dbObj, ...)

\S4method{dbBind}{DomoResult}(res, params, ...)

\S4method{dbFetch}{DomoResult}(res, n = -1, ...)

\S4method{dbGetRowCount}{DomoResult}(res, ...)

\S4method{dbGetStatement}{DomoResult}(res, ...)

\S4method{dbColumnInfo}{DomoResult}(res, ...)

\S4method{dbClearResult}{DomoResult}(res, ...)

\S4method{dbHasCompleted}{DomoResult}(res, ...)
}
\description{
Domo results class
}
\keyword{internal}
